% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adorn_dates}
\alias{adorn_dates}
\title{Function to adorn dates with appropriate formatting}
\usage{
adorn_dates(df, date_var)
}
\description{
Function to adorn dates with appropriate formatting
}
\keyword{internal}
