% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesex_pyramid.R
\name{agesex_pyramid}
\alias{agesex_pyramid}
\title{Age Sex Pyramid}
\usage{
agesex_pyramid(
  df,
  colours = c("#440154", "#fde725"),
  x_breaks = 20,
  y_title = "Proportion of People Percentage (\%)",
  text_size = 12,
  conf_limits = FALSE
)
}
\arguments{
\item{df}{Data frame which has the following columns with names given below
age: char type>> age group can be given as ranges (e.g. "0-4","5-18", "19-64", "65+")
sex: char type>> must be coded as "Female" and "Male"
value: num type>> numerical value used for pyramid (can be proportion or cases)
Optional (set conf_limits parameter to TRUE if error bars are required):
lowercl: num type>> lower confidence limits for value
uppercl: num type>> upper confidence limits for value}

\item{colours}{List of colours for the genders in HEX format in the order Female, Male
Default values for parameter: c("#003B5C", "#007C91")}

\item{x_breaks}{Number of ticks on X axis (default is 20)}

\item{y_title}{Title that  appears on the X axis in as a string}

\item{text_size}{Text size can be modified by providing this number (default set to 15)}

\item{conf_limits}{Boolean variable (True - if error bars are required).
If set to True: the data frame should include the lowercl and uppercl columns}
}
\value{
Returns an age-sex pyramid as a ggplot object
}
\description{
Age Sex Pyramid
}
\examples{
\dontrun{
# Example 1. Simplest use (defaults):
 agesex_pyramid(data_frame)
}
# Example 2. Usage with parameters
\dontrun{
agesex_pyramid(data_frame,
               colours= c("#440154", "#fde725"),
               x_breaks = 20,
               y_title = "Proportion of People Percentage (\%)",
               text_size = 15,
               conf_limits = FALSE)
}
}
