% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agesex_pyramid_grouped.R
\name{agesex_pyramid_grouped}
\alias{agesex_pyramid_grouped}
\title{Age Sex Pyramid using grouped data}
\usage{
agesex_pyramid_grouped(df, colours, x_breaks, y_title, text_size, conf_limits)
}
\arguments{
\item{df}{Data frame which has the following columns with names given below
\describe{
\item{age_group}{Character type. Age group can be given as ranges (e.g. "0-4","5-18", "19-64", "65+")}
\item{sex}{Character type. Must be coded as "Female" and "Male"}
\item{value}{Numeric type. Numerical value used for pyramid (can be proportion or cases)}
\item{lowercl}{Optional. Numeric type. Lower confidence limits for value (required if conf_limits=TRUE)}
\item{uppercl}{Optional. Numeric type. Upper confidence limits for value (required if conf_limits=TRUE)}
}}

\item{colours}{List of colours for the genders in HEX format in the order Female, Male.
Default values: c("#003B5C", "#007C91")}

\item{x_breaks}{Number of ticks on X axis. Default is 20.}

\item{y_title}{Title that appears on the Y axis as a string.}

\item{text_size}{Text size can be modified by providing this number. Default is 15.}

\item{conf_limits}{Boolean variable. If TRUE, error bars are displayed.
If TRUE, the data frame should include the lowercl and uppercl columns.}
}
\value{
Returns an age-sex pyramid as a ggplot object
}
\description{
Age Sex Pyramid using grouped data
}
\examples{
\dontrun{
data <- data.frame(
 age_group = rep(c("0-4", "5-18", "19-64", "65+"), each = 2),
 sex = rep(c("Female", "Male"), times = 4),
 value = c(110, 90, 80, 70, 60, 50, 40, 30)
)

agesex_pyramid_grouped(data, x_breaks=10)
}
}
\keyword{internal}
