% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plotly.R
\name{base_plotly}
\alias{base_plotly}
\title{Creates base plotly object for use across other functions.
Parameters are not passed explicitly to the function, so
function call needs to be proceeded by environment(base_plotly) <- environment()}
\usage{
base_plotly()
}
\value{
plotly object
}
\description{
Creates base plotly object for use across other functions.
Parameters are not passed explicitly to the function, so
function call needs to be proceeded by environment(base_plotly) <- environment()
}
\examples{
\dontrun{
environment(base_plotly) <- environment()
base <- base_plotly()
}
}
\keyword{internal}
