% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_gg_labels}
\alias{clean_gg_labels}
\title{A function to manage 'Ignoring unknown labels:' messages output to
console by ggplot. These happen when items in your_ggplot_object$labels
are NULL or "". The most efficient way of suppressing these is by
dropping these blank labels from the ggplot object before printing it.}
\usage{
clean_gg_labels(p)
}
\arguments{
\item{p}{A ggplot object}
}
\value{
A cleaned ggplot object
}
\description{
A function to manage 'Ignoring unknown labels:' messages output to
console by ggplot. These happen when items in your_ggplot_object$labels
are NULL or "". The most efficient way of suppressing these is by
dropping these blank labels from the ggplot object before printing it.
}
\examples{
\dontrun{
final_ggplot_object <- clean_gg_labels(your_ggplot_object)
}
}
