% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{datebreak_to_d3}
\alias{datebreak_to_d3}
\title{Function to convert ggplot scale_date(date_breaks) values to d3 date
format for use in plotly dtick variable for date axis break setting.}
\usage{
datebreak_to_d3(x)
}
\arguments{
\item{x}{A string giving the distance between breaks like "2 weeks",
or "10 years". Valid specifications are 'sec', 'min', 'hour', 'day',
'week', 'month' or 'year', optionally followed by 's'.
See https://ggplot2.tidyverse.org/reference/scale_date.html}
}
\value{
A string in a format appropriate for plotly dtick date break formatting.
}
\description{
Function to convert ggplot scale_date(date_breaks) values to d3 date
format for use in plotly dtick variable for date axis break setting.
}
\examples{
\dontrun{
datebreak_to_d3("2 months")

datebreak_to_d3("1 week")
}
}
