% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiviz-package.R
\docType{package}
\name{epiviz-package}
\alias{epiviz}
\alias{epiviz-package}
\title{epiviz: Data Visualisation Functions for Epidemiological Data Science Products}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for making epidemiological reporting easier with consistent static and dynamic charts and maps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ukhsa-collaboration/epiviz}
  \item Report bugs at \url{https://github.com/ukhsa-collaboration/epiviz/issues}
}

}
\author{
\strong{Maintainer}: Harshana Liyanage \email{harshana.liyanage@ukhsa.gov.uk} (\href{https://orcid.org/0000-0001-9738-6349}{ORCID})

Authors:
\itemize{
  \item James Duke \email{james.duke@ukhsa.gov.uk}
  \item Alex Bhattacharya \email{alex.bhattacharya@ukhsa.gov.uk} (\href{https://orcid.org/0000-0003-3000-2771}{ORCID})
  \item Liam Fitzpatrick \email{liam.fitzpatrick@ukhsa.gov.uk}
  \item James Leatherland \email{james.leatherland@ukhsa.gov.uk}
  \item Jordan Worth \email{jordan.worth@ukhsa.gov.uk}
  \item Emma Parker \email{emma.k.parker@ukhsa.gov.uk}
}

Other contributors:
\itemize{
  \item Jade Talbot \email{Jade.Talbot@ukhsa.gov.uk} (\href{https://orcid.org/0000-0001-6460-5170}{ORCID}) [contributor]
  \item Karen Crawford \email{Karen.Crawford@ukhsa.gov.uk} [contributor]
  \item Frederick Sloots \email{Frederick.Sloots@ukhsa.gov.uk} [contributor]
  \item Charlie Turner \email{Charlie.Turner@ukhsa.gov.uk} [contributor]
}

}
\keyword{internal}
