% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epiviz.R
\name{epiviz}
\alias{epiviz}
\title{epiviz}
\usage{
epiviz(baseMap = NULL, data, dateRange = NULL, center = NULL, scale = 1,
  ylabel = "Cases/month", width = NULL, height = NULL,
  curveHeight = 0.3, fade = 180, animationDuration = 60, tooltip = NULL,
  drawLegend = TRUE)
}
\description{
Geo-spatial animation of epidemics
}

