% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EpiViz_data2}
\alias{EpiViz_data2}
\title{Example data: Simulated epidemiology results for track 2}
\format{

}
\usage{
EpiViz_data2
}
\description{
A simulated dataset of effect estimates and confidence intervals for 120 variables.
}
\keyword{datasets}
