% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EpiViz_data3}
\alias{EpiViz_data3}
\title{Example data: Mendelian randomization results from an analysis of BMI measured in European women and 123 metabolites}
\format{
A simulated data frame with 123 rows and 8 variables imitating a traditional epidemiology results file from a linear regression:
\describe{
  \item{label}{The name of the outcomes}
  \item{beta}{Effect estimates}
  \item{pvalue}{Pvalues of the effec estimates and standard errors}
  \item{lower_confidence_interval}{Lower confidence interval bound}
  \item{upper_confidence_interval}{Upper confidence interval bound}
  \item{exposure}{The exposure used in the Mendelian randomization alaysis}
  \item{label_units}{The name of the outcomes with units}
  \item{class}{The groups which the outcomes can be put into}
  \item{subclass}{The subgroups which the outcomes can be put into}
  ...
}
}
\usage{
EpiViz_data3
}
\description{
A simulated dataset of linear regression results for 123 outcomes.
}
\keyword{datasets}
