% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Function for determining user's operating system
credit: https://www.r-bloggers.com/2015/06/identifying-the-os-from-r/}
\usage{
get_os()
}
\value{
Name of user's operating system
}
\description{
Function for determining user's operating system
credit: https://www.r-bloggers.com/2015/06/identifying-the-os-from-r/
}
\examples{
\dontrun{
get_os()
}
}
\keyword{internal}
