% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html_bold}
\alias{html_bold}
\title{Function to apply html bold tags to input text strings for use in
plotly font definitions.}
\usage{
html_bold(x)
}
\arguments{
\item{x}{A text string}
}
\value{
A text string wrapped in \if{html}{\out{<b>}} \if{html}{\out{</b>}} for html bold font
}
\description{
Function to apply html bold tags to input text strings for use in
plotly font definitions.
}
\examples{
\dontrun{
html_bold("Chart Title")
}
}
