% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_chart.R
\name{inv_scale_function}
\alias{inv_scale_function}
\title{Function to scale/shift secondary y axis data}
\usage{
inv_scale_function(x, scale, shift)
}
\arguments{
\item{x}{the value to be scaled/shifted}

\item{scale}{the value to scale the data by}

\item{shift}{the value to shift the data by}
}
\value{
scaled/shifted data to plot on secondary y axis
}
\description{
Function to scale/shift secondary y axis data
}
\examples{
inv_scale_function(2,4,5)
}
