% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jsMap.R
\name{jsMap}
\alias{jsMap}
\title{jsMap}
\usage{
jsMap(topojson = NULL, fill = NULL, palette = NULL, mapOptions = NULL,
  map_id = NULL, center = NULL, scale = 500, width = NULL,
  height = NULL)
}
\description{
Create an html map
}

