% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_chart.R
\name{line_chart}
\alias{line_chart}
\title{Line Chart}
\usage{
line_chart(
  dynamic = FALSE,
  base = NULL,
  params = list(df, x, y, ci = NULL, lower = NULL, upper = NULL, error_colour =
    c("#f2c75c"), group_var, line_colour = c("blue"), line_type = "solid", width = 1,
    title = NULL, x_label = NULL, x_label_angle = NULL, y_label = NULL, y_label_angle =
    NULL, y_percent = FALSE, st_theme = NULL, add_points = FALSE, show_gridlines = FALSE,
    show_axislines = TRUE, legend_title = NULL, legend_position = NULL, hline = NULL,
    hline_colour = "red", hline_label = NULL),
  ...
)
}
\arguments{
\item{dynamic}{A logical value. If \code{TRUE}, the line chart will be rendered as a Plotly object for interactivity.
If \code{FALSE}, a static ggplot2 object will be returned.}

\item{base}{A base plotly or ggplot2 object to add the line chart to. Default is \code{NULL}.}

\item{params}{A list containing the following elements:
\itemize{
\item \code{df} A data frame containing the data to be plotted.
\item \code{x} A character string specifying the name of the column in \code{df} to be used for the x-axis.
\item \code{y} A character string specifying the name of the column in \code{df} to be used for the y-axis.
\item \code{group_var} A character string specifying the name of the column in \code{df} to be used for grouping the data.
\item \code{ci} Optional. A character string specifying the column in \code{df} for confidence intervals.
\item \code{lower} Optional. A character string specifying the column in \code{df} for lower bounds of confidence intervals.
\item \code{upper} Optional. A character string specifying the column in \code{df} for upper bounds of confidence intervals.
\item \code{error_colour} The color for error bars. Default is \verb{#f2c75c}.
\item \code{line_colour} List of colours for lines. Default is \code{blue}.
\item \code{line_type} Line type for single graph, or list of line types
Permissable values: "solid", "dotted", "dashed", "longdash", "dotdash"
\item \code{width} A numeric value specifying the width of the lines.
\item \code{title} Optional. A character string specifying the title of the plot.
\item \code{x_label} Optional. A character string specifying the label for the x-axis.
\item \code{x_label_angle} Optional. A numeric value specifying the rotation angle for the x-axis labels.
\item \code{y_label} Optional. A character string specifying the label for the y-axis.
\item \code{y_label_angle} Optional. A numeric value specifying the rotation angle for the y-axis labels.
\item \code{y_percent} Optional. A logical value. If \code{TRUE}, the y-axis will be scaled to percentages.
\item \code{st_theme} Optional. A ggplot2 theme object to customize the style of the plot.
\item \code{add_points} Optional. A logical value. If \code{TRUE}, points will be added to the line chart.
}}

\item{...}{Additional arguments passed to \code{geom_line} for static (ggplot2) plots
or to \code{plot_ly}/\code{add_trace} for dynamic (Plotly) plots, allowing custom
styling of the lines (e.g., \code{alpha}, \code{size}, \code{marker}, etc.).}
}
\value{
A plotly or ggplot2 object representing the line chart.
}
\description{
This function generates a line chart from a data frame using specified x and y variables.
Optionally, the plot can be rendered as an interactive Plotly object. The function also allows
grouping of data based on a specified grouping variable.
}
\examples{
\donttest{
library(dplyr)
library(epiviz)

# Prepare data
test_df <- epiviz::lab_data
test_df$specimen_date <- as.Date(test_df$specimen_date)

summarised_df <- test_df |>
  filter(specimen_date >= as.Date("2023-01-01") &
         specimen_date <= as.Date("2023-12-31")) |>
  group_by(organism_species_name, specimen_date) |>
  summarize(count = n(), .groups = 'drop')

# Create params list
params <- list(
  df = as.data.frame(summarised_df),
  x = "specimen_date",
  y = "count",
  group_var = "organism_species_name",
  line_colour = c("blue", "green", "orange")
)

# Generate static line chart
result <- line_chart(params = params, dynamic = FALSE)
}

}
