% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{London_LA_boundaries_2023}
\alias{London_LA_boundaries_2023}
\title{Local Authority Districts (May 2023) Boundaries for London BGC}
\format{
A dataset with variables:
\describe{
\item{FID}{}
\item{LAD23CD}{Local Authority District code 2023}
\item{LAD23NM}{Local Authority District name 2023}
\item{LAD23NMW}{Local Authority District name 2023 - Welsh}
\item{BNG_E}{}
\item{BNG_N}{}
\item{LONG}{}
\item{LAT}{}
\item{GlobalID}{}
\item{SHAPE_Length}{}
\item{SHAPE_Area}{}
\item{geometry}{}
}
}
\source{
Source: Office for National Statistics licensed under the Open Government Licence v.3.0
Contains OS data © Crown copyright and database right 2023

Data is a subset of the following, including only Local Authority Districts in
Inner London or Outer London.
https://geoportal.statistics.gov.uk/datasets/ons::local-authority-districts-may-2023-boundaries-uk-bgc/explore
}
\usage{
London_LA_boundaries_2023
}
\description{
This data contains the digital vector boundaries for Local Authority Districts,
in London, United Kingdom, as at May 2023. The boundaries are generalised (20m) - clipped to
the coastline (Mean High Water mark). Contains both Ordnance Survey and ONS Intellectual
Property Rights.
}
\keyword{datasets}
