% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nulldefault}
\alias{nulldefault}
\alias{\%||\%}
\title{Null default operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value (default)}
}
\value{
Returns x if not NULL, otherwise returns y
}
\description{
Returns the first argument if it is not NULL, otherwise returns the second argument
}
\details{
Null default operator
}
\examples{
\dontrun{
NULL \%||\% 1  # returns 1
2 \%||\% 1     # returns 2
}
}
\keyword{internal}
