% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{param_assign}
\alias{param_assign}
\title{Function to assign list elements to variables within the parent
environment. Compares input list against a reference vector of expected
list elements, and assigns any that it cannot find a value of 'NULL'.}
\usage{
param_assign(params, reference)
}
\arguments{
\item{params}{List of input parameters.}

\item{reference}{Character vector of expected parameters.}
}
\value{
Assigns parameter values in the parent environment
}
\description{
Function to assign list elements to variables within the parent
environment. Compares input list against a reference vector of expected
list elements, and assigns any that it cannot find a value of 'NULL'.
}
\examples{
\dontrun{
param_assign(params, c("df","x","y","legend","title"))
}
}
