% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{param_rename}
\alias{param_rename}
\title{A function to rename an item in a named list}
\usage{
param_rename(params_list, current_name, new_name)
}
\arguments{
\item{params_list}{A named list}

\item{current_name}{The name of the list item to be modified}

\item{new_name}{The new name of the list item}
}
\value{
A list with the item renamed
}
\description{
A function to rename an item in a named list
}
\examples{
\dontrun{
params <- param_rename(params,"chart_footer_colour","new_name")
}

}
