% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PHEC_boundaries_2016}
\alias{PHEC_boundaries_2016}
\title{PHEC (Dec 2016) Generalised Clipped Boundaries in England}
\format{
A dataset with variables:
\describe{
\item{OBJECTID}{}
\item{phec16cd}{Public Health England Centre code}
\item{phec16nm}{Public Health England Centre name}
\item{bng_e}{}
\item{bng_n}{}
\item{long}{}
\item{lat}{}
\item{GlobalID}{}
\item{SHAPE_Length}{}
\item{SHAPE_Area}{}
\item{geometry}{}
}
}
\source{
Source: Office for National Statistics licensed under the Open Government Licence v.3.0
Contains OS data © Crown copyright and database right 2022
https://open-geography-portalx-ons.hub.arcgis.com/datasets/ons::phec-dec-2016-generalised-clipped-boundaries-in-england/explore
}
\usage{
PHEC_boundaries_2016
}
\description{
This data contains the digital vector boundaries for Public Health England Centres
in England as at December 2016. The boundaries are generalised (20m) - clipped to
the coastline (Mean High Water mark). Contains both Ordnance Survey and ONS Intellectual
Property Rights.
}
\keyword{datasets}
