% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotly_legend_pos}
\alias{plotly_legend_pos}
\title{Function to convert ggplot legend positions into list of
parameters for plotly layout}
\usage{
plotly_legend_pos(x)
}
\arguments{
\item{x}{A ggplot legend position ("top","bootom","right","left")}
}
\value{
A list of parameters for plotly layout.legend
}
\description{
Function to convert ggplot legend positions into list of
parameters for plotly layout
}
\examples{
\dontrun{
plotly_legend_pos("right")
}
}
