% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_chart_plotly.R
\name{point_chart_plotly}
\alias{point_chart_plotly}
\title{Point chart}
\usage{
point_chart_plotly(
  dynamic = FALSE,
  base = NULL,
  params = list(df = NULL, x = NULL, y = NULL, ci = NULL, ci_legend = FALSE,
    ci_legend_title = "Confidence interval", lower = NULL, upper = NULL, error_colours =
    "red", errorbar_width = NULL, group_var = NULL, point_shape = "triangle",
    point_colours = "blue", point_labels = NULL, point_labels_hjust = 0,
    point_labels_vjust = 0, point_labels_nudge_x = 0, point_labels_nudge_y = 0,
    y_sec_axis = FALSE, y_sec_axis_no_shift = FALSE, chart_title = NULL, chart_footer =
    NULL, x_label = NULL, x_label_angle = NULL, y_label = NULL, 
     y_label_angle =
    NULL, y_percent = NULL, st_theme = NULL, x_labels_reverse = FALSE, y_limit_min =
    NULL, y_limit_max = NULL, x_limit_min = NULL, x_limit_max = NULL, x_axis_break_labels
    = NULL, y_axis_break_labels = NULL, x_axis_n_breaks = NULL, y_axis_n_breaks = NULL,
    x_axis_date_breaks = NULL, show_gridlines = FALSE, show_axislines = TRUE,
    legend_title = "", legend_pos = "bottom", hline = NULL, hline_colour = "black",
    hline_width = 0.5, hline_type = "dashed", hline_label = NULL, hline_label_colour =
    "black")
)
}
\arguments{
\item{base}{base}

\item{df}{name of data frame for plotting}

\item{x}{name of x axis variable in data frame}

\item{y}{name of y axis variable in data frame}

\item{group_var}{name of group variable to generate multiple lines
(if required) in data frame}

\item{y_axis}{y_axis}

\item{ci}{indicator for using ribbon or error bar geom (if required),
enter 'errorbar' for error bar, enter 'ribbon' for ribbon}

\item{lower}{lower value for error \ ribbon geom (mandatory if ci argument passed)}

\item{upper}{upper value for error \ ribbon geom (mandatory if ci argument passed)}

\item{error_colours}{if not plotting by group this is the colour of the error
bars or ribbon}

\item{hline}{will display a horizontal line if valid inter passed
provided where multiple lines generated each will be of different type}

\item{y_label}{for provision of an y axis label}

\item{x_label}{for provision of an x axis label}

\item{x_label_angle}{to adjust the x axis label by the degrees of the
integer provided}

\item{y_label_angle}{to adjust the y axis label by the degrees of the
integer provided}

\item{x_labels_reverse}{enter an argument of any value i.e. 'y' to reverse
the x labeling order when using categorical data}

\item{y_min_limit}{set the limit on the y axis scaling by proving an integer}

\item{y_max_limit}{set the limit on the x axis scaling by proving an integer}

\item{x_axis_breaks}{modify the x axis breaks by providing an integer}

\item{legend_pos}{modify the position of the legend (where applicable) with
appropriate value i.e. bottom (default position),
top, right, left}

\item{remove_gridlines}{enter an argument of any value i.e. 'y' to remove
the grid lines}

\item{y_percent}{enter an argument of any value i.e. 'y' to include the \%
symbol for y axis labels}

\item{chart_footer}{enter text for a caption to appear below plot}

\item{labels}{labels to display alongside the plotted geom_points}

\item{labels_hjust}{the horizontal adjustment of the point labels}

\item{labels_vjust}{the vertical adjustment of the point labels}

\item{point_shape}{shape of the points to be plotted (only used when not plotting
by group)}

\item{y_sec_axis_no_shift}{If no shift should be applied to the secondary y-axis}
}
\value{
the final plot
}
\description{
The chart function has 3 mandatory arguments to plot, with additional
arguments as described further below:
}
\examples{

# This example will plot the points on the secondary y-axis with error bars and
# labels showing
# the plotted values that are vertically adjusted from the plotted points
\dontrun{
point_chart(df = plot_df, x = "age", y = "value", ci = "e", upper = "uppercl",
lower = "lowercl", group_var = "ukborn", y_axis = "y2", y_label = "Value",
labels = "value", vjust = -3.5)
}

}
