% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_line_list_for_age_sex_pyramid.R
\name{process_line_list_for_age_sex_pyramid}
\alias{process_line_list_for_age_sex_pyramid}
\title{Process Line List for Age-Sex Pyramid}
\usage{
process_line_list_for_age_sex_pyramid(
  df,
  var_map = list(age = "age", date_of_birth = "date_of_birth", sex = "sex"),
  age_breakpoints = c(0, 5, 19, 65, Inf),
  age_calc_refdate = Sys.Date()
)
}
\arguments{
\item{df}{A data frame containing the line list data.}

\item{var_map}{A list mapping variable names in the data frame to the expected names used in the function.
The list should contain elements:
\describe{
\item{age}{Name of the column containing age values (default is \code{'age'}).}
\item{date_of_birth}{Name of the column containing date of birth values (default is \code{'date_of_birth'}).}
\item{sex}{Name of the column containing sex values (default is \code{'sex'}).
Permitted values for sex include M, F, Male, Female (not case sensitive)}
}}

\item{age_breakpoints}{A numeric vector specifying the breakpoints for age groups. The default is \code{c(0, 5, 19, 65, Inf)}.}

\item{age_calc_refdate}{A Date object specifying the reference date for calculating age from date of birth.
The default is \code{Sys.Date()}.}
}
\value{
A data frame that is aggregated by age group and sex, with columns for age group, sex, value (count),
lower confidence limit, and upper confidence limit.
}
\description{
This function processes a line list data frame, calculating the age from either the provided age column or date of birth.
It filters the data to include only males and females, groups the data into age groups, and summarizes the counts for each age group and sex.
}
\examples{
\dontrun{
  df <- epiviz::lab_data
  processed_df <- process_line_list_for_age_sex_pyramid(df)
}
}
