% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epiviz.R
\name{renderEpiviz}
\alias{renderEpiviz}
\title{Widget render function for use in Shiny}
\usage{
renderEpiviz(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

