% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_function}
\alias{scale_function}
\title{Function to scale/shift secondary y axis labels}
\usage{
scale_function(x, scale, shift)
}
\value{
scaled/shifted y axis values
}
\description{
x the value to be scaled/shifted
scale the value to scale the axis by
shift the value to shift the axis by
}
\examples{
\dontrun{
scale_function(2,4,5)
}
}
