% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_Arial}
\alias{set_Arial}
\title{Function to set chart_font variable to Arial in order to resolve warnings}
\usage{
set_Arial()
}
\value{
chart_font string
}
\description{
Function to set chart_font variable to Arial in order to resolve warnings
}
\examples{
\dontrun{
set_Arial()
}
}
\keyword{internal}
