% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swap_object_names}
\alias{swap_object_names}
\title{A function to swap the names of 2 R objects, used in column_chart() where axis
flipping in plotly to obtain a vertical bar chart involves manually exchanging
all x and y variables.}
\usage{
swap_object_names(name1, name2)
}
\arguments{
\item{name1}{Character, name of the first object}

\item{name2}{Character, name of the second object}
}
\value{
Swaps names of both objects
}
\description{
A function to swap the names of 2 R objects, used in column_chart() where axis
flipping in plotly to obtain a vertical bar chart involves manually exchanging
all x and y variables.
}
\examples{
\dontrun{
dataframe_out <- adorn_date(df = dataframe, date_var = "date_column")
}
}
