% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UK_boundaries_2023}
\alias{UK_boundaries_2023}
\title{Countries (December 2023) Boundaries UK BUC}
\format{
A dataset with variables:
\describe{
\item{FID}{}
\item{CTRY23CD}{Country code 2023}
\item{CTRY23NM}{Country name 2023}
\item{CTRY23NMW}{Country name 2023 - Welsh}
\item{BNG_E}{}
\item{BNG_N}{}
\item{LONG}{}
\item{LAT}{}
\item{GlobalID}{}
\item{SHAPE_Length}{}
\item{SHAPE_Area}{}
\item{geometry}{}
}
}
\source{
Source: Office for National Statistics licensed under the Open Government Licence v.3.0
Contains OS data © Crown copyright and database right 2023
https://geoportal.statistics.gov.uk/datasets/ons::countries-december-2023-boundaries-uk-buc/explore
}
\usage{
UK_boundaries_2023
}
\description{
This data contains the digital vector boundaries for  Countries, in the
United Kingdom, as at December 2023. The boundaries are ultra generalised (500m)
\itemize{
\item clipped to the coastline (Mean High Water mark). Contains both Ordnance Survey
and ONS Intellectual Property Rights.
}
}
\keyword{internal}
