% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_gg.R
\name{base_gg}
\alias{base_gg}
\title{Base ggplot object for line_chart, point_chart, and epi_curve}
\usage{
base_gg()
}
\description{
Internal function to create base ggplot object. Note that the
function call needs to be proceeded by environment(base_gg) <- environment()
to ensure that the function can find variables in the calling environment.
}
\examples{
\dontrun{
environment(base_gg) <- environment()
base <- base_gg()
}
}
\keyword{internal}
