% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_chart.R
\name{col_chart}
\alias{col_chart}
\title{Column Chart}
\usage{
col_chart(
  df,
  base = NULL,
  x = NULL,
  y = NULL,
  group_var = NULL,
  fill = "blue",
  y_axis = "y1",
  position = "dodge",
  ci = NULL,
  lower = NULL,
  upper = NULL,
  error_colour = "red",
  h_line = NULL,
  y_label = NULL,
  x_label = NULL,
  x_label_angle = NULL,
  y_label_angle = NULL,
  x_labels_reverse = NULL,
  y_min_limit = NULL,
  y_max_limit = NULL,
  x_axis_breaks = NULL,
  legend_pos = "bottom",
  remove_gridlines = NULL,
  percent = NULL,
  cap_text = NULL,
  no_shift = FALSE
)
}
\arguments{
\item{df}{the dataframe containing the data to be plotted}

\item{base}{a ggplot instance or NULL if creating a new plot}

\item{x}{the x value column name to be plotted}

\item{y}{the y value column name to be plotted}

\item{group_var}{the variable used to group the bars i.e. region if plotting by region}

\item{fill}{the colour with which to fill the columns}

\item{y_axis}{either "y1" for the primary y-axis or "y2" for the secondary y-axis.}

\item{position}{the positions of the bars to be plotted i.e."dodge", "stack" etc}

\item{ci}{indicator for using ribbon or error bar geom (if required), enter 'e' for error bar, enter any other value for ribbon}

\item{lower}{lower value for error \ ribbon geom (mandatory if ci argument passed)}

\item{upper}{upper upper value for error \ ribbon geom (mandatory if ci argument passed)}

\item{error_colour}{if not plotting by group this is the colour of the error bars or ribbon}

\item{h_line}{will display a horizontal line if valid inter passed}

\item{y_label}{for provision of an y axis label}

\item{x_label}{for provision of an x axis label}

\item{x_label_angle}{to adjust the x axis label by the degrees of the integer provided}

\item{y_label_angle}{to adjust the y axis label by the degrees of the integer provided}

\item{x_labels_reverse}{enter an argument of any value i.e. 'y' to reverse the x labeling order when using categorical data}

\item{y_min_limit}{set the limit on the y axis scaling by proving an integer}

\item{y_max_limit}{set the limit on the x axis scaling by proving an integer}

\item{x_axis_breaks}{modify the x axis breaks by providing an integer}

\item{legend_pos}{modify the position of the legend (where applicable) with appropriate value i.e. bottom (default position), top, right, left}

\item{remove_gridlines}{enter an argument of any value i.e. 'y' to remove the grid lines}

\item{percent}{enter an argument of any value i.e. 'y' to include the \% symbol for y axis labels}

\item{cap_text}{enter text for a caption to appear below plot}

\item{no_shift}{If no shift should be applied to the secondary y-axis}
}
\value{
a ggplot instance
}
\description{
Column Chart
}
\examples{
\dontrun{
library(dplyr)
data <- lab_data |>
 group_by(Organism_Species_Name) |>
 summarise(Count=n())

col_chart(data,
          x="Organism_Species_Name",
          y="Count"
)
}
}
