% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_map.R
\name{epi_map}
\alias{epi_map}
\title{epi_map}
\usage{
epi_map(
  df,
  value_col = NULL,
  inc_shp = TRUE,
  shp_filepath = NULL,
  shp_areacode = NULL,
  data_areacode = NULL,
  lookup = NULL,
  break_intervals = NULL,
  force_cat = TRUE,
  n_breaks = NULL,
  legend_title = NULL,
  legend_pos = "left",
  border_shape_filepath = NULL,
  border_code_col = NULL,
  border_areaname = NULL
)
}
\arguments{
\item{df}{name of data frame for plotting}

\item{value_col}{name of value variable in data frame}

\item{inc_shp}{boolean parameter to indicate whether df includes shapefile data or if this will be uploaded separately.}

\item{shp_filepath}{filepath for the shapefile required for spatial data in function. Note this will not be uploaded if inc_shp = TRUE.}

\item{shp_areacode}{name of areacode variable in shapefile (Mandatory if shp_filepath argument passed).}

\item{data_areacode}{name of areacode variable in dataframe (Mandatory if shp_filepath argument passed).}

\item{lookup}{vector of categories to include in the legend.}

\item{break_intervals}{numeric vector of interval points for the legend (Mandatory if lookup argument passed).}

\item{force_cat}{boolean parameter to determine whether all arguments passed in the lookup argument are used, even if there are no values present in the data.}

\item{n_breaks}{number of break intervals. This argument is an alternative to supplying defined breaks in lookup argument and will
provide a number of evenly distributed breaks as specified. Note if lookup argument is passed this will be ignored.}

\item{legend_title}{string containing the legend title.}

\item{legend_pos}{string containing legend position.}

\item{border_shape_filepath}{filepath for the shapefile if an outer border is required for the map. This should be a higher geography e.g.
if creating an UTLA map- a regional shapefile can be used.}

\item{border_code_col}{name of areacode variable in the border shapefile. Required if a specific area within the border shapefile is required.}

\item{border_areaname}{string containing the name of area required for border shapefile (required if border_code_col argument passed).}
}
\description{
The map function has 2 mandatory arguments to plot, df (data frame containing your data) and value_col (value variable name),
with additional arguments as described further below:
}
\examples{

Examples are provided below for using this function that utilise the in-built DataVis package dummy data to produce
various visualisations, using different arguments.

Example 1: Displays a map for rates per 100,000 across UTLA geographies in London with a border for the outer London region boundaries.
The data frame contains shapefile data which isn't imported separately.  The legend contains specified legend categories with all levels
forcefully included, with a legend title stating "Rate per 100,000" and positioned on the left of the map.
--------------------------------------------------------------------------------------------------------
leg_cats <- c("0.0-0.9 (pre-elimination)", "1.0-4.9 (low incidence)", "5.0-9.9 (low incidence)",
            "10.0-14.9", "15.0-19.9", "20-29.9", "30.0-39.9", ">=40.0")

break_vals <- c( 0.0, 1.0, 5.0, 10.0, 15.0, 20.0, 30.0, 40)


epi_map(df = map,
             inc_shp= TRUE,
             value_col = "value",
             shp_filepath = NULL,
             shp_areacode = NULL,
             data_areacode = NULL,
             lookup = leg_cats,
             break_intervals = break_vals,
             legend_title = "Rate per 100,000",
             force_cat = TRUE,
             n_breaks = NULL,
             border_shape_filepath = "filepath_to_shapefile",
             border_code_col = "RGN21NM",
             border_areaname = "London",
             legend_pos = "left")

Example 2: Displays a map for rates per 100,000 across UTLA geographies in London without a border.
The data frame contains shapefile data which isn't imported separately.  The legend includes 5 breaks which are
not pre-defined, does not include a legend title and is positioned to the right of the map.
--------------------------------------------------------------------------------------------------------
epi_map(df = map,
             inc_shp= TRUE,
             value_col = "value",
             shp_filepath = NULL,
             shp_areacode = NULL,
             data_areacode = NULL,
             lookup = NULL,
             break_intervals = NULL,
             legend_title = NULL,
             force_cat = FALSE,
             n_breaks = 5,
             border_shape_filepath = NULL,
             border_code_col = NULL,
             border_areaname = NULL,
             legend_pos = "right")

}
