% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_intepret.R
\name{llm_interpret}
\alias{llm_interpret}
\title{Interpret Epidemiological Data or Visualisations using LLMs}
\usage{
llm_interpret(input, word_limit = 100, prompt_extension = NULL)
}
\arguments{
\item{input}{An input object which can either be a data frame or a ggplot object.}

\item{word_limit}{The desired word length for the response. Defaults to 100.}

\item{prompt_extension}{Optional additional instructions to extend the standard prompt.}
}
\value{
A narrative or interpretation of the input object as generated by the LLM.
}
\description{
This function interprets a given data frame or ggplot visualization
by sending it to a language model API  via the elmer package.
}
